# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  vulun-scan-backend_internal_models.CreateDomainsRequest:
    properties:
      domains:
        items:
          $ref: '#/definitions/vulun-scan-backend_internal_models.DomainDetail'
        type: array
      organization_id:
        type: integer
    required:
    - domains
    - organization_id
    type: object
  vulun-scan-backend_internal_models.CreateOrganizationRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  vulun-scan-backend_internal_models.CreateSubDomainsRequest:
    properties:
      domain_id:
        type: integer
      sub_domains:
        items:
          type: string
        type: array
    required:
    - domain_id
    - sub_domains
    type: object
  vulun-scan-backend_internal_models.DeleteOrganizationRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  vulun-scan-backend_internal_models.DomainDetail:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  vulun-scan-backend_internal_models.RemoveOrganizationDomainRequest:
    properties:
      domain_id:
        type: integer
      organization_id:
        type: integer
    required:
    - domain_id
    - organization_id
    type: object
  vulun-scan-backend_internal_models.UpdateOrganizationRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  description: 漏洞扫描系统后端 API 文档
  title: Vulun Scan Backend API
  version: "1.0"
paths:
  /domains:
    get:
      description: 支持按ID查询单个域名、按组织ID查询域名列表、分页和排序等
      parameters:
      - description: 域名ID(查询单个域名时使用)
        in: query
        name: id
        type: integer
      - description: 组织ID(查询组织下域名列表时使用)
        in: query
        name: organization_id
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: updated_at
        description: 排序字段
        enum:
        - name
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 域名不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取域名信息(支持多种查询方式)
      tags:
      - 域名管理
  /domains/create:
    post:
      consumes:
      - application/json
      description: 批量创建域名并关联到指定组织
      parameters:
      - description: 域名创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vulun-scan-backend_internal_models.CreateDomainsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 批量创建域名
      tags:
      - 域名管理
  /domains/remove-from-organization:
    post:
      consumes:
      - application/json
      description: 解除指定组织与域名的关联关系，如果域名成为孤儿（没有任何组织关联）则自动删除该域名
      parameters:
      - description: 解除关联请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vulun-scan-backend_internal_models.RemoveOrganizationDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 解除关联成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 组织、域名或关联不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 解除组织与域名的关联
      tags:
      - 域名管理
  /organizations:
    get:
      description: 支持按ID查询单个组织、获取组织列表、分页和排序等
      parameters:
      - description: 组织ID(查询单个组织时使用)
        in: query
        name: id
        type: integer
      - description: 页码,默认1
        in: query
        name: page
        type: integer
      - description: 每页数量,默认10
        in: query
        name: page_size
        type: integer
      - description: '排序字段: id, name, created_at, updated_at,默认updated_at'
        in: query
        name: sort_by
        type: string
      - description: '排序方向: asc, desc,默认desc'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回数据
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 组织不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取组织信息(支持多种查询方式)
      tags:
      - 组织管理
  /organizations/batch-delete:
    post:
      consumes:
      - application/json
      description: 批量删除多个组织及其关联数据，如果组织下有孤儿域名，将一并删除
      parameters:
      - description: 组织ID列表
        in: body
        name: request
        required: true
        schema:
          properties:
            organization_ids:
              items:
                type: integer
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除组织
      tags:
      - 组织管理
  /organizations/create:
    post:
      consumes:
      - application/json
      description: 根据请求体创建一个新组织
      parameters:
      - description: 组织信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vulun-scan-backend_internal_models.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建组织
      tags:
      - 组织管理
  /organizations/delete:
    post:
      consumes:
      - application/json
      description: 删除指定组织及其关联数据
      parameters:
      - description: 删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vulun-scan-backend_internal_models.DeleteOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 组织不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除组织
      tags:
      - 组织管理
  /organizations/update:
    post:
      consumes:
      - application/json
      description: 更新指定组织的名称和描述信息
      parameters:
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vulun-scan-backend_internal_models.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 组织不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新组织
      tags:
      - 组织管理
  /subdomains:
    get:
      description: 支持按ID查询单个子域名、按域名ID筛选、分页查询和排序等
      parameters:
      - description: 子域名ID(查询单个子域名时使用)
        in: query
        name: id
        type: integer
      - description: 域名ID,用于筛选特定域名下的子域名
        in: query
        name: domain_id
        type: integer
      - description: 组织ID,用于筛选特定组织下的子域名
        in: query
        name: organization_id
        type: integer
      - description: 页码,默认1
        in: query
        name: page
        type: integer
      - description: 每页数量,默认10
        in: query
        name: page_size
        type: integer
      - description: '排序字段: id, name, created_at, updated_at,默认updated_at'
        in: query
        name: sort_by
        type: string
      - description: '排序方向: asc, desc,默认desc'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回数据
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 子域名不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取子域名信息(支持多种查询方式)
      tags:
      - 子域名管理
  /subdomains/create:
    post:
      consumes:
      - application/json
      description: 批量创建子域名并关联到指定域名
      parameters:
      - description: 子域名创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vulun-scan-backend_internal_models.CreateSubDomainsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 批量创建子域名
      tags:
      - 子域名管理
swagger: "2.0"